refactor(categories)!: separate admin and public endpoints
- Relocate administrative CRUD operations from CategoryController to the new AdminCategoriesController.
- Set base route to api/admin/categories and enforce FullControlPolicy for all admin methods.
- Establish CategoriesController for public-facing category listing with anonymous access.
- Fix naming collision in unit tests by renaming test classes and updating constructor references.
- Update integration tests and Postman E2E collections to match the new /admin routing structure.

BREAKING CHANGE: Category admin endpoints are moved from api/category to api/admin/categories. The public list is now accessible via api/category.
