+++ CookingBlogBackend/PostApiService.Tests/Helper/HttpHelper.cs
@@ -7,7 +7,8 @@
             public static class Categories
             {
                 public const string Base = "/api/category";
-                public const string GetById = "/api/category/{0}";
+                public const string AdminBase = "/api/admin/categories";
+                public const string GetById = "/api/admin/categories/{0}";
             }
 
             public static class Posts
