+++ postman/collection.json
@@ -193,13 +193,14 @@
               }
             },
             "url": {
-              "raw": "{{baseUrl}}/api/Category",
+              "raw": "{{baseUrl}}/api/admin/categories",
               "host": [
                 "{{baseUrl}}"
               ],
               "path": [
                 "api",
-                "Category"
+                "admin",
+                "categories"
               ]
             }
           },
@@ -236,13 +237,14 @@
             "method": "GET",
             "header": [],
             "url": {
-              "raw": "{{baseUrl}}/api/Category/{{temp_category_id}}",
+              "raw": "{{baseUrl}}/api/admin/categories/{{temp_category_id}}",
               "host": [
                 "{{baseUrl}}"
               ],
               "path": [
                 "api",
-                "Category",
+                "admin",
+                "categories",
                 "{{temp_category_id}}"
               ]
             }
@@ -341,13 +343,14 @@
               }
             },
             "url": {
-              "raw": "{{baseUrl}}/api/Category/{{temp_category_id}}",
+              "raw": "{{baseUrl}}/api/admin/categories/{{temp_category_id}}",
               "host": [
                 "{{baseUrl}}"
               ],
               "path": [
                 "api",
-                "Category",
+                "admin",
+                "categories",
                 "{{temp_category_id}}"
               ]
             }
@@ -989,13 +992,14 @@
             "method": "DELETE",
             "header": [],
             "url": {
-              "raw": "{{baseUrl}}/api/Category/{{temp_category_id}}",
+              "raw": "{{baseUrl}}/api/admin/categories/{{temp_category_id}}",
               "host": [
                 "{{baseUrl}}"
               ],
               "path": [
                 "api",
-                "Category",
+                "admin",
+                "categories",
                 "{{temp_category_id}}"
               ]
             }
